Utility: Add a 'profile' parameter to the avatar command
Partially completes #9

It lets the user view a 'user' or 'server' avatar.
The default is the guild avatar, then the user avatar,
and lastly the default one generated by Discord.

Command errors for the 'profile' parameter are thrown only if the user
selects the 'server' profile and the target doesn't have a guild avatar.

Also default to the 'gif' format if available for the target avatar.